improve code for math and accuracy

+++ app/src/main/java/com/gregor/lauritz/healthdashboard/data/local/entity/WorkoutRecordEntity.kt
@@ -22,5 +22,5 @@ data class WorkoutRecordEntity(
     val zone4Minutes: Float,
     val zone5Minutes: Float,
     val trimp: Float,
-    val avgHr: Int,
+    val avgHr: Float,
 )
